Better log API changes to Mono.Android.dll
If (when?) an API change is made to Mono.Android.dll,
`src/Mono.Android/ApiCompatLinesToAdd.txt` will be created containing
the lines that need to be added to e.g.
`tests/api-compatibility/acceptable-breakages-vReference-*.txt`
so that the changes are ignored.

These lines can then be easily copied into
`acceptable-breakages-vReference-*.txt` if deemed useful.

Previously, *all* changes were printed to the build log, which was
annoying to deal with because (1) there may be duplicates, and
(2) the lines would contain `TaskId`/etc. "noise" that would need to
be removed in order to be used.