[Mono.Android] pass in `Context` at startup
Calling `Android.App.Application.Context` currently reads the
`mono.MonoPackageManager.Context` field, which is set on startup by a
`ContentProvider`. `mono.MonoPackageManager` does not exist in
NativeAOT (or CoreCLR), so let's approach this differently.

Each `ContentProvider` can pass in the context to `JNIEnvInit` on
startup, then the first call to `Android.App.Application.Context` does
not have to read a field any longer as we already have the handle in
managed code.