[java-runtime] move `MonoPackageManager.Context` to `ApplicationRegistration`
Calling `Android.App.Application.Context` currently reads the
`mono.MonoPackageManager.Context` field, which is set on startup by a
`ContentProvider`. `mono.MonoPackageManager` does not exist in
NativeAOT (or CoreCLR), so let's move the field to:

* `net.dot.android.ApplicationRegistration.Context`

The `ApplicationRegistration` class is already used in NativeAOT, so
this can work for all of the runtimes.