[Mono.Android-Tests] Remove obsolete `.shproj` file. (#9758)
`Mono.Android-Tests` still uses a `.shproj` file from the time when we had both a Classic XA and .NET version.  However, VS does not seem to support this anymore, and none of the included files appear in the IDE, making it hard to write new tests.

Remove the `.shproj` file and update the `.csproj` file to directly point to needed files.

Also do some other cleanups like removing tests that are no longer run and fixing directory structure.

Verified that both the `main` version and this PR run 885 tests on CI.