[Xamarin.Android.Build.Tasks] use `mono.android.jar` only for `$(AndroidCodegenTarget)=XAJavaInterop1`
`mono.android.jar/dex` are compiled with
`$(AndroidCodegenTarget)=XAJavaInterop1`, and so this causes a problem
if you build a project targeting `$(AndroidCodegenTarget)=JavaInterop1`
as incorrect Java stubs will be used.

For example, `dotnet new maui` using NativAOT we can get a crash such as:

    02-14 16:23:29.884 19260 19260 E .hellonativeaot: No implementation found for void mono.android.Runtime.register(java.lang.String, java.lang.Class, java.lang.String) (tried Java_mono_android_Runtime_register and Java_mono_android_Runtime_register__Ljava_lang_String_2Ljava_lang_Class_2Ljava_lang_String_2) - is the library loaded, e.g. System.loadLibrary?

Which is caused by `mono.android.Runtime.register` being called by
`mono/android/view/View_OnClickListenerImplementor.java` which is
compiled with `$(AndroidCodegenTarget)=XAJavaInterop1`.

To fix this:

* Several places already check if `$(_AndroidUseMarshalMethods)=true`
  and skip usage of `mono.android.jar`.

* Add an additional check to verify `$(AndroidCodegenTarget)=XAJavaInterop1`

Previously, `$(AndroidCodegenTarget)` was also allowed to blank and
default to `XAJavaInterop1`, but it seems better to explicitly set it.

I also updated the `NativAOT()` test to verify
`mono/android/views/View_OnClickListenerImplementor.java` is generated
on disk.