Fixity Fix
dotnet/java-interop#1302 fixed the
`Could not determine Java type corresponding to System.Byte[]` crash.

What replaced it was:

    Mono.Android.NET_Tests, Java.InteropTests.JnienvTest.NewObjectArrayWithNonJavaType / Release
    System.ArgumentException : Could not determine Java type corresponding to System.Type, System.Private.CoreLib, Version=10.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e. Arg_ParamName_Name, type

The deal is that [`JavaNativeTypeManager.ToJniName(Type)`][0]
defaults to using `java/lang/Object` if there is no type mapping.

This is "reasonable" because of the "auto-box into
`Android.Runtime.JavaObject`" behavior, meaning it *is*
possible to create a Java array that "holds" `System.Type` instances.
(Because the Java-side array is actually a `java.lang.Object[]`, each
instance of which is an `Android.Runtime.JavaObject`, which in turn
holds the `System.Type` instance.)

However, this is *not* a semantic that
`JniRuntime.JniTypeManager.GetTypeSignature()` maintains, and thus
this was lost in 1b1f1452f6b05707418d6605c06e106e6a2a6381.

Update `JNIEnv.FindClass(Type)` so that if `.GetTypeSignature()`
can't find the typemapping, default to `java/lang/Object`.

This allows all tests to pass, locally.

[0]: https://github.com/dotnet/java-interop/blob/62635a3ffee4c9ec421eb029b86e61267a544f92/src/Java.Interop.Tools.TypeNameMappings/Java.Interop.Tools.TypeNameMappings/JavaNativeTypeManager.cs#L164-L168