Update System.IO.Compression to use the best compression.
`System.IO.Compression`` uses the default compression level which is not
the best compression level. This change updates the compression
level to the best compression level since the size of the Apk is
very important.

Unfortunately, the best compression enum level is not available on `netstandard2.0`.
So we have to use a cast to get it to work.