[Xamarin.Android.Build.Tasks] Error APT2265 for invalid Asset Path (#9809)
Context: https://dev.azure.com/devdiv/DevDiv/_workitems/edit/2356711

Users regularly run into the following error.

	error APT2000: The system cannot find the specified file. (2):
	C:\Users\xxx\source\repos\AdventureWorks.MAUI\AdventureWorks.MAUI\obj\Debug\net9.0-android\assets

One of the reasons behind this is [an `aapt2` bug][0] in which it
cannot traverse directories containing non-ASCII characters.
This has been a long standing bug that does not seem to be getting
fixed any time soon. 

As we know why this issue occurs, lets at least raise the specific
error code with instructions on how to fix it. 

[0]: https://issuetracker.google.com/issues/188679588