[jit-times] Use a single `long` instead of `(long, int, int)` (#9484)
The operator+ and operator- operators for `Timestamp` type can produce
invalid Timestamps.  All of Timestamp parts should be either
(0 or positive) or (0 or negative), but statements like
`result.milliseconds--` don't take this value into consideration.

The reporting is also broken: the tool parses `methods.txt` in the
assumption that it is produced by a single-threaded app, and it seems
that it's not the case in reality.  This is why many `self` timings
are negative.

Instead of treating a timestamp as a "tuple" of
`(long seconds, int milliseconds, int nanoseconds)`, treat it as a
single `long nanoseconds`.  This reduces the amount of range of times
that can be parsed, but the new reduced upper limit of 2262-04-11
gives us around 238 years of not needing to care, while massively
simplifying the logic and fixing bugs.