[CoreCLR] More CLR hosting support (#9572)
Implements but does not complete CoreCLR  host support.
The runtime build will be enabled later, which will also be used to
fix and update tests, potentially introducing new ones.  This is in
the interest to keep this PR as small as possible and getting the
changes that affect the whole build system into `main` as quickly as
possible.