Fix Debugger timeout on 32 bit devices.
Context https://github.com/dotnet/android/issues/9573

Not sure why, but we are getting an odd problem were the
`timeout` value we pass to mono for the debugger is not
long enough. This is causing the debugger to timeout.

What we see is `timeout=3` rather than `timeout=1740583188`.
Lets upgrade to use `std:format` to see if that helps as it
with automatically take care of the conversion for us.