Try to prevent the SetupOSSpecificPropertyes from running when not needed
The target sets up environment for the NativeAOT ILC compiler and it requires to find llvm-objcopy
and objcopy (among others) utilities in the PATH. We have code that sets PATH up to point to the NDK
bin directory, but it runs too late to prevent errors when the setup target doesn't find all the
required binaries. It appears the target runs whether or not NativeAOT is used, and is dependent on
the NativeCompilationDuringPublish property being `true`, which is the default value, assigned to it
whenever ilccompiler targets are included in the build (which seems to be our case no matter which
android runtime we target in the given build).

Let's see if this commit helps.