[nativeaot] fix `.resx` files in apps
The `EmbeddedResources_ShouldBeLocalized()` test was failing under a
NativeAOT context:

    Embedded string resource did not contain expected value after changing CultureInfo.
    String lengths are both 1. Strings differ at index 0.
    Expected: "b"
    But was:  "a"
    -----------^

After reviewing a `.binlog`, it appears the `_ComputeIlcCompileInputs`
MSBuild target was not working as expected:

https://github.com/dotnet/runtime/blob/cc803458ab4dabf020b462873be5bf56f1640b1e/src/coreclr/nativeaot/BuildIntegration/Microsoft.NETCore.Native.Publish.targets#L3-L11

`@(IntermediateSatelliteAssembliesWithTargetPath)` was blank!

The problem being the way Android has an "inner" build per RID:

* `@(IntermediateSatelliteAssembliesWithTargetPath)` is set in the
  "outer" build.

* `@(IntermediateSatelliteAssembliesWithTargetPath)` is blank in the
  "inner" build per RID.

* The "inner" build is where the ILC compiler runs, and so it doesn't
  see the current project's `*.resource.dll` file.

To fix this, we can pass `$(_OuterIntermediateSatelliteAssembliesWithTargetPath)`
into the inner build.

I updated an MSBuild test to run normally and in a NativeAOT context.