[Xamarin.Android.Build.Tasks] make "managed typemap" runtime agnostic
For NativeAOT, we implemented a "managed" typemap that is trimmer-safe.

In order to test its performance characteristics, we can make this
typemap useable for Mono and CoreCLR as well:

* Move `NativeAotTypeManager`, `NativeAotValueManager`, and
  `TypeMapping` types to `Mono.Android.dll`

* Rename `NativeAot*` to `Managed*`

* Add a new private `$(_AndroidTypeMapImplementation)` MSBuild
  property that can be set to `llvm-ir` or `managed`.

* Add a new trimmer feature flag
  `Android.Runtime.RuntimeFeature.ManagedTypeMap` that when `true`
  uses the managed typemap on any runtime.

I added a test that verifies `dotnet run` succeeds for all typemap
implementations.

Note that NativeAOT will *only* support the managed typemap.