﻿[XABT] Scan for JLO needed wrappers in `LinkAssembliesNoShrink`. (#9893)
Today, application builds use Cecil to scan through all Android assemblies at least twice:

- Debug: `_LinkAssembliesNoShrink`, `_GenerateJavaStubs`
- Release: `_ILLink`, `_GenerateJavaStubs`, `_RemoveRegisterAttribute`

This is a costly operation that we would like to only perform once to help speed up application builds.  The long-term goal is to move the (many) steps currently performed by `_GenerateJavaStubs` into "linker" steps.

This PR moves the Cecil scanning required for the first step: "generating Java Callable Wrappers".

## Implementation

This does not move generating JCW Java code to the linker, it only moves scanning for the JLO information needed to generate JCWs to the linker.  This information is persisted as an XML file in `/obj` next to the processed assembly.

Example:

- `obj/Debug/net10.0-android/android/assets/MyApplication.dll`
- `obj/Debug/net10.0-android/android/assets/MyApplication.jlo.xml`

Doing it this way has two benefits: it helps keep the linker from getting too complex (eg: generating Java code) and it helps incremental builds.  If an assembly has not changed since the last build, the JLO information in the `.jlo.xml` is still valid and will be reused.  The assembly does not need to be re-scanned.  (For example, the ~70 AndroidX assemblies that the MAUI template uses do not need to be re-scanned on every Debug build.)

The process of actually generating the JCWs is done by a new `GenerateJavaCallableWrappers` task in the `_GenerateJavaStubs` target that consumes the `.jlo.xml` files and outputs the `.java` files.

## Release Builds

In an ideal Release build, we would (probably?) run this new step as part of the `ILLink` invocation.  However there are some tricky limitations to running in the linker pipeline, chiefly being that the linker cannot load external assemblies like `Java.Interop.Tool.JavaCallableWrappers.dll` (and dependencies) that we need.  This can be worked around by including the needed source files in the linker assembly, but this will require additional work.

Instead, for now, we will add an invocation a new `_LinkAssembliesAdditionalSteps` target to linked builds that runs after `ILLink`. This calls the existing `LinkAssembliesNoShrink` task to scan for JLOs but does not execute the other linker tasks that have already been run. This allows us to move forward with converting to "linker steps" for making Debug builds faster without the extra work required to move these steps into `ILLink`.  A future enhancement can perform this work if desired.  (There is also [some question](https://github.com/dotnet/runtime/issues/107211) as to whether we actually want to create new `ILLink` linker steps.)

Note this temporarily leaves the `$(_AndroidJLOCheckedBuild)` flag that compares the generated Java files to the previous ones.  This will likely be useful ensuring future moved steps are correct as well.