[ci] Don't re-run `dotnet test` if there are no auto retry tests. (#9940)
Context: d00b779a5b5f405af967d85e28e570e873631a6b

Commit d00b779a added a unit test retry mechanism to address flaky
emulator tests.

This works by:

  - Running the unit tests via `dotnet test`
  - Running `dotnet-test-slicer` to create a test filter that only
    contains the failed tests
  - Rerunning `dotnet test` with the filtered test list

In the case where there are no failed tests, a "dummy" filter is
created that does not match any tests.

Most of the time this is fine and the `dotnet test` invocation with
the dummy filter takes a couple of seconds to determine it has
nothing to do and exits.  However, this ***no-op*** can ***itself***
hang, causing the build to hang for ~3 hours and eventually fail.  🤦 

Instead, add a check to see if the failed test list is empty and
fully skip the `dotnet test` invocation if it is, avoiding the
potential hang.

Result:

> ▶️ **run MSBuildDeviceIntegration On Device - macOS-1 (Auto-Retry)**
>
> Skipping step due to condition evaluation.

![image](https://github.com/user-attachments/assets/acb84d32-f2d5-4e56-bf63-8290c4777f04)