[build] see if we can allow long branch names
If you open a PR with a long branch name, it can error with:

    NuGet.Build.Tasks.Pack.targets(221,5): error NU5123: Warning As Error: The file 'package/services/metadata/core-properties/d9cd2ffb8cfc41bdbfe90e7dbc7228fb.psmdcp' path, name, or both are too long. Your package might not work without long file path support. Please shorten the file path or file name.

Let's set `$(WarningAsMessages)=NU5123` for pull requests.

I think the warning, in general, is good to know if we create a path
inside a `.nupkg` that is too long. But for PRs, we should ignore it.

This also reverts some commits that "shorten" branch names, which
should no longer be needed:

* 7651b64895ecf2812e85cd570b61f3950b3db09c
* 6d9b295652ef5292427cb15b2afce469c6ed940f