[XABT] Break Cecil usage out of `TypeMapGenerator`. (#9996)
In preparation for future changes to the typemap generation process, this commit refactors the usage of `Mono.Cecil` out of the `TypeMapGenerator` class.  (Part of the "move `GenerateJavaStubs` to linker steps" effort.)

This facilitates a future state where:
- The `Module[Debug|Release]Data` entries are created from `Mono.Cecil` data in the `LinkAssembliesNoShrink`/`ILLink` step and persisted to disk.
- The `GenerateTypeMappings` task reads the persisted information to generate the typemap.

This refactor could be done as part of that future change, but seemed like a smaller change that could be more easily reviewed and committed separately.