[XABT] Remove support for `$(_AndroidGenerateNativeAssembly)` = `false` (typemaps). (#10001)
Context: #9292

In #9292, we removed support for the "instant run" feature. This feature required the use of typemaps that existed outside of the "native" assemblies ($(_AndroidGenerateNativeAssembly) = false). This support is no longer needed, and in fact causes a build error if it is used:

Xamarin.Android.Common.targets(2132,3): error XA3006: Could not compile native assembly file: typemaps.x86_64.ll
As such, clean up our codebase to remove support for this.

Additionally, unused fields in the Module[Debug|Release]Data and TypeMap[Debug|Release]Entry data classes were removed to help make it clearer what data is required to generate typemaps.