[build] Use SharedFramework.Sdk to create workload packs (#10000)
Our workload packaging logic has been updated to use more of the
[Microsoft.DotNet.SharedFramework.Sdk][0] MSBuild SDK included in the
dotnet/arcade repo.

This SDK allows us to easily exclude the symbols files that we've
been shipping directly in all of our workload packs, and instead move
them into `*.symbols.nupkg` packages that can be archived and
uploaded to a symbol server separately.

Note: the `_GenerateFrameworkList` target is in [sharedfx.targets][1].

[0]: https://github.com/dotnet/arcade/tree/21f46f494265f11e124c47ee868fac199768a35b/src/Microsoft.DotNet.SharedFramework.Sdk
[1]: https://github.com/dotnet/arcade/blob/2b1931c05dd6ceb89120131a8b39eaa81735179a/src/Microsoft.DotNet.SharedFramework.Sdk/targets/sharedfx.targets#L398-L429