Try to preserve the app cache on uninstall.
Context #9259

Due to various reasons `adb` may not be able to preserve the app cache
on uninstall. To work around this we have to use the `-k` flag to
uninstall the app when using the `cmd package uninstall` call.
This will keep the app data and cache on the device.

However this might also fail. So in that case we need to completely
uninstall the app and then install it again.