Make the NativeAOT debugger script more flexible (#10016)
* Make the NativeAOT debugger script more flexible

Make it able to detect the OS-specific path to the NDK toolchain binaries, as well
as the path to the NDK root directory using a number of environment variables used
at one time or another for this purpose.

Instead of hard-coding path to the `lldb-server` binary, which is located in a
versioned directory, use `clang` itself to print full path to the binary, thus
making the script LLVM version agnostic.

* Maybe it works here too?

* Use prefixed clang binary

* Fixlets

* Add quotes to CLANG_PREFIX values

---------

Co-authored-by: Dean Ellis <dellis1972@googlemail.com>