[XABT] GeneratePackageManagerJava - Convert `NativeCodeGenState` to Cecil-less DTOs. (#10058)
Context: https://github.com/dotnet/android/issues/10062

_One_ issue with our desire to move around LLVM Marshal Method build code so that it can work with Ready2Run is that our current implementation relies on holding Cecil `AssemblyDefinition` objects open in global state across multiple targets.  While these objects are open, no other process can access our assemblies, making it harder to reorder targets and tasks.

As a first step towards refactoring this process, let's move the data into a DTO that does not require Cecil.  Additionally, this makes it clear _what_ data is actually consumed by `<GeneratePackageManagerJava>` rather than "entire assemblies" so we can focus on how to properly obtain and persist the data.

This still relies on `BuildEngine4.RegisterTaskObject` to store the data, but a likely future step would be to persist it to a file instead.