[ci] run MSBuild "SmokeTests" category on all runtimes
For each "smoke test" that build an Android application project
(libraries not really relevant), we run the test on all runtimes.

For each test, I generally added two more test cases for NativeAOT and
CoreCLR.

This is a first step to ensure that we don't break CoreCLR or
NativeAOT support at build time.