[native] Common infrastructure for native shared library linking (#9990)
Context: https://github.com/dotnet/android/pull/9938

PR #9938 is an attempt to build `libnet-android.Debug.so` and
`libnet-android.Release.so` so that CoreCLR is *statically linked* in
to `libnet-android*.so`.  This reduces the number of native libraries
that must be placed into the `.apk`, and allows the `.apk` to be
smaller, as the native linker can remove unused code.

*Note*: CoreCLR is ***not*** yet statically linked into
`libnet-android*.so`.  This is *preliminary* work to support that.

This iteration of static linking also requires that *all* symbols be
resolvable, by requiring `lld --no-undefined` by default.  This in
turn requires that we ship *copies* of `libc.so` et al, so that *all*
symbols needed by `libnet-android*.so` can be resolved, including
those from the Android NDK.

Update `build-tools/xaprepare` to extract required native libraries
from the NDK, and add those to the .NET for Android Workload packs.

Update `Xamarin.Android.Build.Tasks.dll` to cleanup and refactor
`ld` invocation, so that `libnet-android*.so` is built with
`ld --no-undefined`.  This makes it easier to maintain and update the
linker code in the future, as well as ensures that we use consistent
linking flags across the board.

Updates the LLVM IR code generator bits, also extracted from #9938,
which aren't strictly related to native linking, but they are placed
here in order to make #9938 smaller.