[XABT] Separate marshal method storage from `MarshalMethodClassifier` to `MarshalMethodCollection`. (#10077)
Today the marshal method scanning process is tightly coupled to scanning for JavaCallableWrappers. That is, when building `CallableWrapperType` objects, the `MarshalMethodsClassifier` is used to classify methods as marshal methods by calling `ShouldBeDynamicallyRegistered (...)`.

However `ShouldBeDynamicallyRegistered` doesn't just return a simple `bool`, it has the side effect of building a list of `MarshalMethodEntry` objects that are used extensively throughout the build process.

We are likely going to need to build this list in places other than JavaCallableWrappers scanning: assembly rewriting will likely be before JCWs, and `GenerateNativeMarshalMethodSources` will likely be after JCWs.

As such:
- Rework `MarshalMethodsClassifier` to simply classify methods as LLVM marshal method eligible or not (dynamically registered).
- Move storage to a new `MarshalMethodsCollection`.

Additionally, rework `MarshalMethodsClassifier` to also be able to detect marshal methods that have already been rewritten and provide the information later steps (`GenerateJavaStubs`, `GenerateNativeMarshalMethodSources`) need to function properly.  This isn't in use yet, but will be needed in the future.