[Mono.Android] fix "exception unboxing" for `ManagedValueManager` (#10100)
Context: https://github.com/dotnet/android/pull/10095

After enabling `Java.Interop-Tests.dll` for CoreCLR/NativeAOT,
`Java.InteropTests.JavaExceptionTests.InnerExceptionIsNotAProxy()`
was failing with:

	Expected: same as <System.InvalidOperationException: Managed Exception!>
	But was:  <Java.Interop.JavaProxyThrowable: System.InvalidOperationException: Managed Exception!

This is because `AndroidRuntime.cs` was doing:

	if (JNIEnvInit.ValueManager is AndroidValueManager vm) {
	    return vm.UnboxException (value);
	}

And in this case `ManagedValueManager` was being used, and so the
exception would not be "unboxed".

We can fix this by using `JNIEnvInit.ValueManager?.PeekValue()`
instead and the test now passes.

I also noticed that `AndroidValueManager` had code in
`TryUnboxPeerObject()` for:

	var proxy = value as Android.Runtime.JavaProxyThrowable;
	if (proxy != null) {
	    result  = proxy.InnerException;
	    return true;
	}
	return base.TryUnboxPeerObject (value, out result);

While the base class handled `Java.Interop.JavaProxyThrowable`.
This was missing for `ManagedValueManager`, so I added it.